Move utils into one unified folder
