+++ utils/Phrase.ts
